REFACTOR: moved credential login into a separate module
